# frontend/openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "Myle vl2 API",
    "version": "0.1.0"
  },
  "paths": {
...
    "/api/v1/auth/dev-login": {
      "post": {
        "tags": [
          "auth"
        ],
        "summary": "Dev Login",
        "operationId": "dev_login_api_v1_auth_dev_login_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/DevLoginRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/DevLoginResponse"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/auth/login": {
      "post": {
        "tags": [
          "auth"
        ],
        "summary": "Login With Password",
        "description": "Email + password; user must have ``hashed_password`` set (see migrations / admin tooling).",
        "operationId": "login_with_password_api_v1_auth_login_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/LoginRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/DevLoginResponse"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
...
    "/api/v1/follow-ups": {
      "get": {
        "tags": [
          "follow-ups"
        ],
        "summary": "List Follow Ups",
        "operationId": "list_follow_ups_api_v1_follow_ups_get",
        "parameters": [
          {
            "name": "limit",
            "in": "query",
            "required": false,
            "schema": {
              "type": "integer",
              "maximum": 100,
              "minimum": 1,
              "default": 50,
              "title": "Limit"
            }
          },
          {
            "name": "offset",
            "in": "query",
            "required": false,
            "schema": {
              "type": "integer",
              "minimum": 0,
              "default": 0,
              "title": "Offset"
            }
          },
          {
            "name": "open_only",
            "in": "query",
            "required": false,
            "schema": {
              "type": "boolean",
              "description": "If true, only rows with completed_at IS NULL",
              "default": true,
              "title": "Open Only"
            },
            "description": "If true, only rows with completed_at IS NULL"
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/FollowUpListResponse"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      },
      "post": {
        "tags": [
          "follow-ups"
        ],
        "summary": "Create Follow Up",
        "operationId": "create_follow_up_api_v1_follow_ups_post",
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/FollowUpCreate"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/FollowUpPublic"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/follow-ups/{follow_up_id}": {
      "patch": {
        "tags": [
          "follow-ups"
        ],
        "summary": "Update Follow Up",
        "operationId": "update_follow_up_api_v1_follow_ups__follow_up_id__patch",
        "parameters": [
          {
            "name": "follow_up_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "title": "Follow Up Id"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/FollowUpUpdate"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/FollowUpPublic"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      },
      "delete": {
        "tags": [
          "follow-ups"
        ],
        "summary": "Delete Follow Up",
        "operationId": "delete_follow_up_api_v1_follow_ups__follow_up_id__delete",
        "parameters": [
          {
            "name": "follow_up_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "title": "Follow Up Id"
            }
          }
        ],
        "responses": {
          "204": {
            "description": "Successful Response"
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
...
  },
  "components": {
    "schemas": {
...
      "DevLoginRequest": {
        "properties": {
          "role": {
            "type": "string",
            "enum": [
              "admin",
              "leader",
              "team"
            ],
            "title": "Role"
          }
        },
        "type": "object",
        "required": [
          "role"
        ],
        "title": "DevLoginRequest"
      },
      "DevLoginResponse": {
        "properties": {
          "ok": {
            "type": "boolean",
            "title": "Ok",
            "default": true
          }
        },
        "type": "object",
        "title": "DevLoginResponse"
      },
      "FollowUpCreate": {
        "properties": {
          "lead_id": {
            "type": "integer",
            "minimum": 1.0,
            "title": "Lead Id"
          },
          "note": {
            "type": "string",
            "maxLength": 2000,
            "minLength": 1,
            "title": "Note"
          },
          "due_at": {
            "anyOf": [
              {
                "type": "string",
                "format": "date-time"
              },
              {
                "type": "null"
              }
            ],
            "title": "Due At"
          }
        },
        "type": "object",
        "required": [
          "lead_id",
          "note"
        ],
        "title": "FollowUpCreate"
      },
      "FollowUpListResponse": {
        "properties": {
          "items": {
            "items": {
              "$ref": "#/components/schemas/FollowUpPublic"
            },
            "type": "array",
            "title": "Items"
          },
          "total": {
            "type": "integer",
            "title": "Total"
          },
          "limit": {
            "type": "integer",
            "title": "Limit"
          },
          "offset": {
            "type": "integer",
            "title": "Offset"
          }
        },
        "type": "object",
        "required": [
          "items",
          "total",
          "limit",
          "offset"
        ],
        "title": "FollowUpListResponse"
      },
      "FollowUpPublic": {
        "properties": {
          "id": {
            "type": "integer",
            "title": "Id"
          },
          "lead_id": {
            "type": "integer",
            "title": "Lead Id"
          },
          "lead_name": {
            "type": "string",
            "title": "Lead Name",
            "description": "Parent lead name at read time"
          },
          "note": {
            "type": "string",
            "title": "Note"
          },
          "due_at": {
            "anyOf": [
              {
                "type": "string",
                "format": "date-time"
              },
              {
                "type": "null"
              }
            ],
            "title": "Due At"
          },
          "completed_at": {
            "anyOf": [
              {
                "type": "string",
                "format": "date-time"
              },
              {
                "type": "null"
              }
            ],
            "title": "Completed At"
          },
          "created_by_user_id": {
            "type": "integer",
            "title": "Created By User Id"
          },
          "created_at": {
            "type": "string",
            "format": "date-time",
            "title": "Created At"
          }
        },
        "type": "object",
        "required": [
          "id",
          "lead_id",
          "lead_name",
          "note",
          "created_by_user_id",
          "created_at"
        ],
        "title": "FollowUpPublic"
      },
      "FollowUpUpdate": {
        "properties": {
          "note": {
            "anyOf": [
              {
                "type": "string",
                "maxLength": 2000,
                "minLength": 1
              },
              {
                "type": "null"
              }
            ],
            "title": "Note"
          },
          "due_at": {
            "anyOf": [
              {
                "type": "string",
                "format": "date-time"
              },
              {
                "type": "null"
              }
            ],
            "title": "Due At",
            "description": "Omit to leave unchanged; null clears due date"
          },
          "completed": {
            "anyOf": [
              {
                "type": "boolean"
              },
              {
                "type": "null"
              }
            ],
            "title": "Completed",
            "description": "True = mark done now; False = reopen"
          }
        },
        "type": "object",
        "title": "FollowUpUpdate"
      },
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            },
            "type": "array",
            "title": "Detail"
          }
        },
        "type": "object",
        "title": "HTTPValidationError"
      },
...
      "LoginRequest": {
        "properties": {
          "email": {
            "type": "string",
            "maxLength": 320,
            "minLength": 3,
            "title": "Email"
          },
          "password": {
            "type": "string",
            "maxLength": 512,
            "minLength": 1,
            "title": "Password"
          }
        },
        "type": "object",
        "required": [
          "email",
          "password"
        ],
        "title": "LoginRequest",
        "description": "Email kept as str so dev domains like ``@myle.local`` validate without DNS rules."
      },
...
      "ValidationError": {
        "properties": {
          "loc": {
            "items": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "integer"
                }
              ]
            },
            "type": "array",
            "title": "Location"
          },
          "msg": {
            "type": "string",
            "title": "Message"
          },
          "type": {
            "type": "string",
            "title": "Error Type"
          },
          "input": {
            "title": "Input"
          },
          "ctx": {
            "type": "object",
            "title": "Context"
          }
        },
        "type": "object",
        "required": [
          "loc",
          "msg",
          "type"
        ],
        "title": "ValidationError"
      },
...
    }
  }
}